fix: rename hooks folder
