+++ src/components/ConcentratedLiquidity/AddLiquidity/FeeSelector/index.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Box, Button, Text } from 'src/components';
 import { useChainId } from 'src/hooks';
-import { useFeeTierDistributionHook, usePoolsHook } from 'src/hooks/concentratedLiquidity/hooks';
-import { PoolState } from 'src/hooks/concentratedLiquidity/hooks/types';
+import { useFeeTierDistributionHook, usePoolsHook } from 'src/hooks/elixir/hooks';
+import { PoolState } from 'src/hooks/elixir/hooks/types';
 import usePrevious from 'src/hooks/usePrevious';
 import FeeOption from './FeeOption';
 import { FeeTierPercentageBadge } from './FeeTierPercentageBadge';
