+++ src/components/ConcentratedLiquidity/DetailModal/EarnWidget/Remove/index.tsx
@@ -7,8 +7,8 @@ import { useChainId, useLibrary, usePangolinWeb3 } from 'src/hooks';
 import { MixPanelEvents } from 'src/hooks/mixpanel';
 import useTransactionDeadline from 'src/hooks/useTransactionDeadline';
 import { useWalletModalToggle } from 'src/state/papplication/hooks';
-import { useElixirRemoveLiquidityHook } from 'src/state/pburn/concentratedLiquidity';
-import { useDerivedBurnInfo } from 'src/state/pburn/concentratedLiquidity/common';
+import { useElixirRemoveLiquidityHook } from 'src/state/pburn/elixir';
+import { useDerivedBurnInfo } from 'src/state/pburn/elixir/common';
 import { useUserSlippageTolerance } from 'src/state/puser/hooks';
 import { ButtonWrapper, RemoveLiquidityWrapper, RemoveWrapper } from './styles';
 import { RemoveProps } from './types';
