+++ src/components/ConcentratedLiquidity/DetailModal/EarnWidget/RemoveDrawer/types.tsx
@@ -1,4 +1,4 @@
-import { PositionDetails } from 'src/state/pwallet/concentratedLiquidity/types';
+import { PositionDetails } from 'src/state/pwallet/elixir/types';
 
 export interface RemoveDrawerProps {
   isOpen: boolean;
