+++ src/components/ConcentratedLiquidity/DetailModal/EarnWidget/index.tsx
@@ -4,10 +4,10 @@ import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Box, Button, Loader, Stat, Text, TransactionCompleted } from 'src/components';
 import { useChainId, useLibrary, usePangolinWeb3 } from 'src/hooks';
-import { useConcLiqPositionFeesHook } from 'src/hooks/concentratedLiquidity/hooks';
-import { usePool } from 'src/hooks/concentratedLiquidity/hooks/common';
+import { useConcLiqPositionFeesHook } from 'src/hooks/elixir/hooks';
+import { usePool } from 'src/hooks/elixir/hooks/common';
 import { MixPanelEvents } from 'src/hooks/mixpanel';
-import { useElixirCollectEarnedFeesHook } from 'src/state/pwallet/concentratedLiquidity/hooks';
+import { useElixirCollectEarnedFeesHook } from 'src/state/pwallet/elixir/hooks';
 import { unwrappedToken } from 'src/utils/wrappedCurrency';
 import RemoveDrawer from './RemoveDrawer';
 import { ClaimWrapper, RewardWrapper, Root, StatWrapper } from './styles';
