+++ src/components/ConcentratedLiquidity/DetailModal/IncreasePosition/index.tsx
@@ -11,11 +11,11 @@ import { MixPanelEvents } from 'src/hooks/mixpanel';
 import useTransactionDeadline from 'src/hooks/useTransactionDeadline';
 import { useUSDCPriceHook } from 'src/hooks/useUSDCPrice';
 import { useWalletModalToggle } from 'src/state/papplication/hooks';
-import { Field } from 'src/state/pmint/concentratedLiquidity/atom';
-import { useDerivedMintInfo, useMintActionHandlers, useMintState } from 'src/state/pmint/concentratedLiquidity/hooks';
+import { Field } from 'src/state/pmint/elixir/atom';
+import { useDerivedMintInfo, useMintActionHandlers, useMintState } from 'src/state/pmint/elixir/hooks';
 import { useUserSlippageTolerance } from 'src/state/puser/hooks';
-import { useElixirAddLiquidityHook } from 'src/state/pwallet/concentratedLiquidity/hooks';
-import { useDerivedPositionInfo } from 'src/state/pwallet/concentratedLiquidity/hooks/evm';
+import { useElixirAddLiquidityHook } from 'src/state/pwallet/elixir/hooks';
+import { useDerivedPositionInfo } from 'src/state/pwallet/elixir/hooks/evm';
 import { useCurrencyBalance } from 'src/state/pwallet/hooks/common';
 import { unwrappedToken, wrappedCurrency } from 'src/utils/wrappedCurrency';
 import OutofRangeWarning from '../../OutofRangeWarning';
