+++ src/components/ConcentratedLiquidity/DetailModal/IncreasePosition/types.tsx
@@ -1,4 +1,4 @@
-import { PositionDetails } from 'src/state/pwallet/concentratedLiquidity/types';
+import { PositionDetails } from 'src/state/pwallet/elixir/types';
 
 export type IncreasePositionProps = {
   position?: PositionDetails;
