+++ src/components/ConcentratedLiquidity/DetailModal/Tabs/DetailTab/detailTab.stories.tsx
@@ -2,7 +2,7 @@ import { CHAINS, ChainId, FeeAmount, Token } from '@pangolindex/sdk';
 import { ComponentStory } from '@storybook/react';
 import { BigNumber } from 'ethers';
 import React from 'react';
-import { PositionDetails } from 'src/state/pwallet/concentratedLiquidity/types';
+import { PositionDetails } from 'src/state/pwallet/elixir/types';
 import { DetailTabProps } from './types';
 import DetailTab from '.';
