+++ src/components/ConcentratedLiquidity/index.tsx
@@ -4,9 +4,9 @@ import { Box, Button, Loader, Text } from 'src/components';
 import { BIG_INT_ZERO } from 'src/constants';
 import { useChainId } from 'src/hooks';
 import useDebounce from 'src/hooks/useDebounce';
-import { useMintActionHandlers } from 'src/state/pmint/concentratedLiquidity/hooks';
-import { useGetUserPositionsHook } from 'src/state/pwallet/concentratedLiquidity/hooks';
-import { PositionDetails } from 'src/state/pwallet/concentratedLiquidity/types';
+import { useMintActionHandlers } from 'src/state/pmint/elixir/hooks';
+import { useGetUserPositionsHook } from 'src/state/pwallet/elixir/hooks';
+import { PositionDetails } from 'src/state/pwallet/elixir/types';
 import { Visible } from 'src/theme/components';
 import AddLiquidity from './AddLiquidity';
 import DetailModal from './DetailModal';
