+++ src/hooks/elixir/hooks/evm.ts
@@ -16,7 +16,7 @@ import { BigNumber } from 'ethers';
 import { useEffect, useMemo, useState } from 'react';
 import { ElixirPoolType, ElixirTick, useElixirPools } from 'src/apollo/elixirPools';
 import { useFeeTierDistributionQuery } from 'src/apollo/feeTierDistribution';
-import { CONCENTRATE_POOL_STATE_INTERFACE } from 'src/constants/abis/concentratedPool';
+import { CONCENTRATE_POOL_STATE_INTERFACE } from 'src/constants/abis/elixirPool';
 import { useChainId } from 'src/hooks';
 import { TokenReturnType } from 'src/hooks/tokens/constant';
 import { useConcLiqNFTPositionManagerContract, useTokenContract } from 'src/hooks/useContract';
