+++ src/state/pmint/elixir/hooks.ts
@@ -22,8 +22,8 @@ import { useTranslation } from 'react-i18next';
 import { BIG_INT_ZERO } from 'src/constants';
 import { PairState, usePair } from 'src/data/Reserves';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
-import { usePool } from 'src/hooks/concentratedLiquidity/hooks/common';
-import { PoolState } from 'src/hooks/concentratedLiquidity/hooks/types';
+import { usePool } from 'src/hooks/elixir/hooks/common';
+import { PoolState } from 'src/hooks/elixir/hooks/types';
 import { useCurrency } from 'src/hooks/useCurrency';
 import { tryParseAmount } from 'src/state/pswap/hooks/common';
 import { wrappedCurrency, wrappedCurrencyAmount } from 'src/utils/wrappedCurrency';
