+++ src/state/pwallet/elixir/hooks/evm.ts
@@ -12,11 +12,11 @@ import { BigNumber } from 'ethers';
 import { useMemo } from 'react';
 import { BIPS_BASE } from 'src/constants/swap';
 import { useChainId, useLibrary, usePangolinWeb3 } from 'src/hooks';
-import { usePool } from 'src/hooks/concentratedLiquidity/hooks/common';
+import { usePool } from 'src/hooks/elixir/hooks/common';
 import { useTokensHook } from 'src/hooks/tokens';
 import { useConcLiqNFTPositionManagerContract as useElixirNFTPositionManagerContract } from 'src/hooks/useContract';
 import { useCurrency } from 'src/hooks/useCurrency';
-import { Field } from 'src/state/pmint/concentratedLiquidity/atom';
+import { Field } from 'src/state/pmint/elixir/atom';
 import { useSingleCallResult, useSingleContractMultipleData } from 'src/state/pmulticall/hooks';
 import { useTransactionAdder } from 'src/state/ptransactions/hooks';
 import { calculateGasMargin, waitForTransaction } from 'src/utils';
