+++ src/utils/computeSurroundingTicks.ts
@@ -1,6 +1,6 @@
 import { JSBI, Token, tickToPrice } from '@pangolindex/sdk';
 import { Ticks } from 'src/apollo/allTicks';
-import { TickProcessed } from 'src/hooks/concentratedLiquidity/chart/types';
+import { TickProcessed } from 'src/hooks/elixir/chart/types';
 
 const PRICE_FIXED_DIGITS = 8;
